:hammer: Apply some refactoring
